# backend/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Beatris API
  version: 1.0.0
  description: API веб-приложения медицинского центра эстетической медицины Beatris
servers:
  - url: /api/v1
    description: Основной относительный адрес API
...
paths:
...
  /service-categories:
    get:
      tags:
        - Services
      summary: Получить категории услуг
      operationId: getServiceCategories
      responses:
        "200":
          description: Категории услуг успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceCategory"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services:
    get:
      tags:
        - Services
      summary: Получить список услуг
      operationId: getServices
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: categorySlug
          in: query
          description: Slug категории услуги
          required: false
          schema:
            type: string
            example: injection-cosmetology
        - name: minPrice
          in: query
          description: Минимальная стоимость услуги
          required: false
          schema:
            type: number
            minimum: 0
            example: 10000
        - name: maxPrice
          in: query
          description: Максимальная стоимость услуги
          required: false
          schema:
            type: number
            minimum: 0
            example: 50000
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Список услуг успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /services/{slug}:
    get:
      tags:
        - Services
      summary: Получить услугу по slug
      operationId: getServiceBySlug
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Услуга успешно получена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /appointments:
    post:
      tags:
        - Appointments
      summary: Создать запись или заявку на приём
      description: Endpoint доступен гостю и авторизованному пациенту. Если пользователь авторизован, данные пациента могут быть подставлены из профиля.
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Запись на приём успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /product-categories:
    get:
      tags:
        - Products
      summary: Получить категории товаров
      operationId: getProductCategories
      responses:
        "200":
          description: Категории товаров успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductCategory"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products:
    get:
      tags:
        - Products
      summary: Получить список товаров
      operationId: getProducts
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: categorySlug
          in: query
          description: Slug категории товара
          required: false
          schema:
            type: string
            example: skincare
        - name: minPrice
          in: query
          description: Минимальная цена товара
          required: false
          schema:
            type: number
            minimum: 0
            example: 5000
        - name: maxPrice
          in: query
          description: Максимальная цена товара
          required: false
          schema:
            type: number
            minimum: 0
            example: 30000
        - name: sort
          in: query
          description: Сортировка товаров
          required: false
          schema:
            type: string
            enum:
              - price_asc
              - price_desc
              - newest
              - popular
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Список товаров успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/{slug}:
    get:
      tags:
        - Products
      summary: Получить товар по slug
      operationId: getProductBySlug
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Товар успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /orders:
    post:
      tags:
        - Orders
      summary: Оформить заказ
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Заказ успешно оформлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/my:
    get:
      tags:
        - Orders
      summary: Получить мои заказы
      operationId: getMyOrders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Заказы пользователя успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Получить заказ по id
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Заказ успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/{id}/payment:
    post:
      tags:
        - Orders
      summary: Создать платёж для заказа
      operationId: createOrderPayment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Платёж успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentCreateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /recommendations/my:
    get:
      tags:
        - Recommendations
      summary: Получить мои рекомендации
      operationId: getMyRecommendations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Рекомендации успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /recommendations:
    post:
      tags:
        - Recommendations
      summary: Создать рекомендацию для пациента
      operationId: createRecommendation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationRequest"
      responses:
        "201":
          description: Рекомендация успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /doctor/schedule:
    get:
      tags:
        - Doctor
      summary: Получить расписание текущего врача
      operationId: getDoctorSchedule
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-05-20"
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-05-27"
      responses:
        "200":
          description: Расписание врача успешно получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DoctorScheduleItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /doctor/appointments:
    get:
      tags:
        - Doctor
      summary: Получить записи пациентов текущего врача
      operationId: getDoctorAppointments
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-05-20"
      responses:
        "200":
          description: Записи пациентов успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /admin/orders/{id}:
    patch:
      tags:
        - Admin
      summary: Изменить статус заказа
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Статус заказа успешно изменён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token, передаваемый в заголовке Authorization
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Refresh token, хранящийся в защищённой HttpOnly cookie
  parameters:
    Id:
      name: id
      in: path
      description: Уникальный идентификатор ресурса
      required: true
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
    Slug:
      name: slug
      in: path
      description: Человекочитаемый идентификатор ресурса
      required: true
      schema:
        type: string
        example: biorevitalizatsiya
    Search:
      name: search
      in: query
      description: Поисковая строка
      required: false
      schema:
        type: string
        example: акне
    Page:
      name: page
      in: query
      description: Номер страницы
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Количество элементов на странице
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 400
            message: Некорректные данные запроса
            error: Bad Request
    Unauthorized:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 401
            message: Необходимо выполнить вход в систему
            error: Unauthorized
    Forbidden:
      description: Недостаточно прав доступа
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 403
            message: Недостаточно прав для выполнения действия
            error: Forbidden
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 404
            message: Запрашиваемый ресурс не найден
            error: Not Found
    ValidationError:
      description: Ошибка валидации данных
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            statusCode: 422
            message:
              - phone должен быть указан
              - serviceId должен быть UUID
            error: Unprocessable Entity
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 500
            message: Внутренняя ошибка сервера
            error: Internal Server Error
  schemas:
...
    AppointmentStatus:
      type: string
      enum:
        - pending
        - confirmed
        - completed
        - cancelled
      example: confirmed
...
    ServiceCategory:
      type: object
      required:
        - id
        - title
        - slug
      properties:
        id:
          type: string
          format: uuid
          example: 2c0b07ae-4b62-4c5f-b4bc-4edb886725e2
        title:
          type: string
          example: Инъекционная косметология
        slug:
          type: string
          example: injection-cosmetology
    Service:
      type: object
      description: Услуга медицинского центра
      required:
        - id
        - title
        - slug
        - description
        - price
        - category
      properties:
        id:
          type: string
          format: uuid
          example: 661b9091-86a4-4512-aa0d-162c8d0c2db0
        title:
          type: string
          example: Биоревитализация
        slug:
          type: string
          example: biorevitalizatsiya
        description:
          type: string
          example: Инъекционная процедура для увлажнения и восстановления кожи
        price:
          type: number
          minimum: 0
          example: 25000
        durationMinutes:
          type: integer
          nullable: true
          example: 60
        category:
          $ref: "#/components/schemas/ServiceCategory"
        imageUrl:
          type: string
          nullable: true
          example: /uploads/services/biorevitalizatsiya.jpg
        contraindications:
          type: string
          nullable: true
          example: Беременность, воспалительные процессы, индивидуальные противопоказания
...
    ServiceListResponse:
      type: object
      required:
        - items
        - meta
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    CreateAppointmentRequest:
      type: object
      required:
        - serviceId
        - specialistId
        - date
        - time
        - patientName
        - patientPhone
      properties:
        serviceId:
          type: string
          format: uuid
        specialistId:
          type: string
          format: uuid
        date:
          type: string
          format: date
          example: "2026-05-20"
        time:
          type: string
          example: "14:30"
        patientName:
          type: string
          example: Айгуль Нурланова
        patientPhone:
          type: string
          example: "+77001234567"
        comment:
          type: string
          nullable: true
          example: Хочу записаться на первичную консультацию
    Appointment:
      type: object
      description: Запись на приём
      required:
        - id
        - service
        - specialist
        - date
        - time
        - status
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
          nullable: true
        patientName:
          type: string
          example: Айгуль Нурланова
        patientPhone:
          type: string
          example: "+77001234567"
        service:
          $ref: "#/components/schemas/Service"
        specialist:
          $ref: "#/components/schemas/Specialist"
        date:
          type: string
          format: date
          example: "2026-05-20"
        time:
          type: string
          example: "14:30"
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        comment:
          type: string
          nullable: true
          example: Первичная консультация
        patientContactStatus:
          type: string
          enum:
            - not_contacted
            - contacted
            - agreed
            - declined
          example: agreed
        patientContactedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-09T10:00:00Z"
        patientContactComment:
          type: string
          nullable: true
          example: Пациент согласовал перенос на 15:30
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentStatusHistoryItem"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
...
    ProductCategory:
      type: object
      required:
        - id
        - title
        - slug
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Средства для ухода
        slug:
          type: string
          example: skincare
    Product:
      type: object
      description: Товар интернет-магазина
      required:
        - id
        - title
        - slug
        - description
        - price
        - category
        - stock
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Крем для восстановления кожи
        slug:
          type: string
          example: vosstanavlivayushchiy-krem
        description:
          type: string
          example: Профессиональное средство для домашнего ухода
        price:
          type: number
          minimum: 0
          example: 18000
        category:
          $ref: "#/components/schemas/ProductCategory"
        imageUrl:
          type: string
          nullable: true
          example: /uploads/products/cream.jpg
        stock:
          type: integer
          minimum: 0
          example: 12
...
    ProductListResponse:
      type: object
      required:
        - items
        - meta
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    CreateOrderRequest:
      type: object
      required:
        - deliveryAddress
        - recipientName
        - recipientPhone
      properties:
        deliveryAddress:
          type: string
          example: г. Атырау, ул. Абая, 10
        recipientName:
          type: string
          example: Айгуль Нурланова
        recipientPhone:
          type: string
          example: "+77001234567"
        comment:
          type: string
          nullable: true
          example: Позвонить перед доставкой
...
    Order:
      type: object
      description: Заказ пользователя
      required:
        - id
        - items
        - totalPrice
        - paymentStatus
        - orderStatus
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: number
          example: 36000
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        orderStatus:
          $ref: "#/components/schemas/OrderStatus"
        deliveryAddress:
          type: string
          example: г. Атырау, ул. Абая, 10
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
    PaymentCreateResponse:
      type: object
      required:
        - paymentUrl
        - paymentId
      properties:
        paymentId:
          type: string
          format: uuid
        paymentUrl:
          type: string
          example: https://payments.example.com/pay/123
        expiresAt:
          type: string
          format: date-time
          example: "2026-05-11T10:15:00Z"
    Recommendation:
      type: object
      description: Рекомендация врача после приёма
      required:
        - id
        - patientId
        - doctorId
        - text
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
          nullable: true
        text:
          type: string
          example: Использовать SPF-средство утром и увлажняющий крем вечером
        productIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
          example: "2026-05-11T10:00:00Z"
    CreateRecommendationRequest:
      type: object
      required:
        - patientId
        - text
      properties:
        patientId:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
          nullable: true
        text:
          type: string
          example: Использовать SPF-средство утром и увлажняющий крем вечером
        productIds:
          type: array
          items:
            type: string
            format: uuid
    DoctorScheduleItem:
      type: object
      required:
        - date
        - startTime
        - endTime
        - isAvailable
      properties:
        date:
          type: string
          format: date
          example: "2026-05-20"
        startTime:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "18:00"
        isAvailable:
          type: boolean
          example: true
    UpdateOrderStatusRequest:
      type: object
      required:
        - orderStatus
      properties:
        orderStatus:
          $ref: "#/components/schemas/OrderStatus"
...